api_server/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1
requests==2.31.0
celery==5.3.6
pika==1.3.2
ldap3==2.9.1
ntlm-auth==1.5.0
pyspnego==0.10.2
